+++ src/agdb/collections/vec.rs
@@ -1,3 +1,5 @@
 pub mod old_storage_vec_iterator;
-pub mod vec_storage_iterator_dynamic_sized;
-pub mod vec_storage_iterator_fixed_size;
+pub mod vec_dynamic_sized;
+pub mod vec_dynamic_sized_iterator;
+pub mod vec_fixed_sized;
+pub mod vec_fixed_sized_iterator;
